Added a tableview with contentfiltering to OrderView. Added method getAllRentals() to OrderViewController. Changed properties in Rental.java, Inventory.java, Film.java, Language.java and Staff.java. References to foreign keys was changed from references to the foreign entity's id to a Datatype of that entity. Ex. Rental.java no longer references to a short customerID, it now references to a Customer customer joined by the column customer_id.

+++ src/main/java/org/edusystems/entities/Language.java
@@ -10,7 +10,7 @@ public class Language {
     @Id
     @Column(name = "language_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int languageId;
+    private short languageId;
 
     @Column(name = "name")
     private String name;
